Renaming PhpUnit Adapter
